feat: separate infra and functional tests to ensure no repeated tests
