+++ auth/feature-tests/infrastructure/vitest-step-definitions/GovUKMobilePostAuthenticationFunction.steps.ts
