+++ auth/feature-tests/infrastructure/vitest-step-definitions/GovUKMobileWebApplicationFirewall.steps.ts
@@ -5,7 +5,7 @@ import { readFileSync } from "fs";
 import { load } from "js-yaml";
 
 const feature = await loadFeature(
-  "feature-tests/vitest-features/GovUKMobileWebApplicationFirewall.feature"
+  "feature-tests/infrastructure/vitest-features/GovUKMobileWebApplicationFirewall.feature"
 );
 
 let template: Template;
